+++ src/main/kotlin/core/model/InventoryDTO.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.SerialName
 
 @Serializable
 data class InventoryDTO(
-    @SerialName("id") val id: Int = 0,
-    @SerialName("c") val configId: Int = 0,
+    @SerialName("id") val id: Long = 0,
+    @SerialName("c") val configId: Long = 0,
     @SerialName("b") val bags: List<InventoryItemBagDTO>? = null,
 )
