+++ src/main/kotlin/core/model/config/ConfigBalanceDTO.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ConfigBalanceDTO(
     @SerialName("key") val key: String = "",
-    @SerialName("val") val value: Int = 0,
+    @SerialName("val") val value: Long = 0,
 )
