+++ src/main/kotlin/core/model/config/ConfigCityWheelDTO.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ConfigCityWheelDTO(
     @SerialName("cb") val wheelBuffs: List<ConfigCityWheelBuffDTO> = emptyList(),
-    @SerialName("e") val listOfEventIds: List<Int> = emptyList(),
+    @SerialName("e") val listOfEventIds: List<Long> = emptyList(),
 )
