+++ src/main/kotlin/core/model/config/ConfigDurationVo.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfigDurationVo(
-    @SerialName("d") val duration: Int = 0,
+    @SerialName("d") val duration: Long = 0,
     @SerialName("tu") val timeUnit: String = "",
 )
