+++ src/main/kotlin/core/model/config/ConfigPlayfieldItemImprovementSlotDTO.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfigPlayfieldItemImprovementSlotDTO(
-    @SerialName("sid") val id: Int = 0,
+    @SerialName("sid") val id: Long = 0,
     @SerialName("oid") val orderId: Int = 0,
 )
