+++ src/main/kotlin/core/model/config/ConfigPlayfieldItemPermissionDTO.kt
@@ -5,10 +5,10 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfigPlayfieldItemPermissionDTO(
-    @SerialName("id") val id: Int = 0,
+    @SerialName("id") val id: Long = 0,
     @SerialName("p") val initialPermissions: Int = 0,
     @SerialName("bc") val buyCap: Int = 0,
     @SerialName("c") val realCurrencyCost: Int = 0,
-    @SerialName("m") val modifier: Int = 0,
-    @SerialName("e") val eventIdList: List<Int> = emptyList(),
+    @SerialName("m") val modifier: Long = 0,
+    @SerialName("e") val eventIdList: List<Long> = emptyList(),
 )
