+++ src/main/kotlin/core/model/config/ConfigProfessionalSpecializationDTO.kt
@@ -5,8 +5,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfigProfessionalSpecializationDTO(
-    @SerialName("cpsi") val id: Int = 0,
-    @SerialName("dt") val departmentTagId: Int = 0,
+    @SerialName("cpsi") val id: Long = 0,
+    @SerialName("dt") val departmentTagId: Long = 0,
     @SerialName("r") val rank: Int = 0,
     @SerialName("v") val slotMaxValue: Int = 0,
     @SerialName("gfx") val gfxId: Int = 0,
