+++ src/main/kotlin/core/model/config/ConfigProfessionalTraitDTO.kt
@@ -5,8 +5,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfigProfessionalTraitDTO(
-    @SerialName("v") val value: Int = 0,
-    @SerialName("cpt") val id: Int = 0,
+    @SerialName("v") val value: Long = 0,
+    @SerialName("cpt") val id: Long = 0,
     @SerialName("lo") val localeid: Int = 0,
     @SerialName("gfx") val gfxid: Int = 0,
 )
