+++ src/main/kotlin/core/model/config/ConfigQuestDTO.kt
@@ -1,12 +1,13 @@
 package core.model.config
 
+import core.model.quest.QuestRequirementDTO
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
 data class ConfigQuestDTO(
-    @SerialName("et") val timeLeft: Int = 0,
+    @SerialName("et") val timeLeft: Long = 0,
     @SerialName("s") val resumable: Boolean = false,
-    @SerialName("cgId") val questGiverGFXid: Int = 0,
+    @SerialName("cgId") val questGiverGFXid: Long = 0,
     @SerialName("rq") val questRequirements: List<QuestRequirementDTO> = emptyList(),
 )
