+++ src/main/kotlin/core/model/quest/QuestCostVo.kt
@@ -8,8 +8,8 @@ data class QuestCostVo (
     // explicit check of [QuestSystemRewardTypeConstants]
     @SerialName("t") val rewardType: String = QuestSystemRewardTypeConstants.QUEST,
     // explicit amount checking in `QuestFactory.getQuestRewardDto`
-    @SerialName("a") val amount: Int = 1,
-    @SerialName("id") val rewardConfigID: Int = 1,
+    @SerialName("a") val amount: Long = 1,
+    @SerialName("id") val rewardConfigID: Long = 1,
 ) {
     init {
         require(rewardType in QuestSystemRewardTypeConstants.ALL) {
