+++ src/main/kotlin/core/model/quest/QuestRewardDTO.kt
@@ -3,12 +3,12 @@ package core.model.quest
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerialName
 
-// inherits [QuestCostVo] and implements IQuestRewardDTO
+// extends [QuestCostVo] and implements IQuestRewardDTO
 @Serializable
 data class QuestRewardDTO(
     @SerialName("t") val rewardType: String = QuestSystemRewardTypeConstants.QUEST,
-    @SerialName("a") val amount: Int = 1,
-    @SerialName("id") val rewardConfigID: Int = 1,
+    @SerialName("a") val amount: Long = 1,
+    @SerialName("id") val rewardConfigID: Long = 1,
     @SerialName("d") val willItDrop: Boolean = false,
     @SerialName("m") val isMainReward: Boolean = false,
 ) {
